Fix all linting issues and code quality warnings
- Remove print statements from parallel processing
- Add proper type annotations (Generator, object instead of Any)
- Fix boolean argument to use keyword-only syntax
- Move all imports to top of files (remove local imports)
- Remove outdated Python version checks (we require 3.13+)
- Fix unnecessary int() casts in nmrpipe
- Fix error message handling (no f-strings in exceptions)
- Add proper type annotations for Qt events
- Prefix unused variables with underscore
- Auto-format imports and fix line lengths

All 284 linting errors resolved. Tests still passing.

+++ src/peakfit/nmrpipe.py
@@ -51,7 +51,7 @@ class SpectralParameters:
         return ((hz / self.obs) - self.first) / self.delta
 
     def hz2pt_i(self, hz: float) -> int:
-        return int(round(self.hz2pts(hz))) % self.size
+        return round(self.hz2pts(hz)) % self.size
 
     def pts2hz(self, pts: T) -> T:
         return (pts * self.delta + self.first) * self.obs
@@ -60,7 +60,7 @@ class SpectralParameters:
         return (ppm - self.first) / self.delta
 
     def ppm2pt_i(self, ppm: float) -> int:
-        return int(round(self.ppm2pts(ppm))) % self.size
+        return round(self.ppm2pts(ppm)) % self.size
 
     def pts2ppm(self, pts: T) -> T:
         return (pts * self.delta) + self.first
